+++ src/backend_skia/mod.rs
@@ -74,6 +74,22 @@ impl OutputImage for skia::Surface {
     fn make_vec(&mut self) -> Vec<u8> {
         self.data().to_vec()
     }
+
+    fn make_rgba_vec(&mut self) -> Vec<u8> {
+        use rgb::FromSlice;
+        use std::mem::swap;
+
+        let mut data = self.make_vec();
+
+        // BGRA_Premultiplied -> BGRA
+        svgfilters::demultiply_alpha(data.as_bgra_mut());
+        // BGRA -> RGBA.
+        if skia::Surface::is_bgra() {
+            data.as_bgra_mut().iter_mut().for_each(|p| swap(&mut p.r, &mut p.b));
+        }
+
+        data
+    }
 }
 
 /// Renders SVG to image.